ui/mantine-ui: share duration-expression explain notes and fix dedupe key
collectNotes and durationExprNote were duplicated verbatim in
ExplainView.tsx and Selector.tsx. The min_of/max_of dedupe key
interpolated formatDurationNode (which returns a ReactNode) into a
string, so both operands stringified to [object Object] and distinct
expressions collapsed to a single note.

Extract both helpers into a shared durationExprNotes module and build
the dedupe key from the string serializeDurationNode instead.
