+++ web/ui/mantine-ui/src/pages/query/ExplainViews/ExplainView.tsx
@@ -1,6 +1,6 @@
-import { FC, ReactNode } from "react";
-import { Alert, Text, Anchor, Card, Divider, List } from "@mantine/core";
-import ASTNode, { nodeType, DurationNode } from "../../../promql/ast";
+import { FC } from "react";
+import { Alert, Text, Anchor, Card, Divider } from "@mantine/core";
+import ASTNode, { nodeType } from "../../../promql/ast";
 // import AggregationExplainView from "./Aggregation";
 // import BinaryExprExplainView from "./BinaryExpr/BinaryExpr";
 // import SelectorExplainView from "./Selector";
@@ -8,89 +8,12 @@ import funcDocs from "../../../promql/functionDocs";
 import { escapeString } from "../../../promql/utils";
 import { formatPrometheusDuration } from "../../../lib/formatTime";
 import { formatDurationNode } from "../../../promql/format";
+import { durationExprNote } from "../../../promql/durationExprNotes";
 import classes from "./ExplainView.module.css";
 import SelectorExplainView from "./Selector";
 import AggregationExplainView from "./Aggregation";
 import BinaryExprExplainView from "./BinaryExpr/BinaryExpr";
 import { IconInfoCircle } from "@tabler/icons-react";
-const collectNotes = (
-  node: DurationNode | null,
-  notes: ReactNode[],
-  seen: Set<string>
-): void => {
-  if (!node || node.type === "numberLiteral") {
-    return;
-  }
-  if (node.op === "step") {
-    if (!seen.has("step")) {
-      seen.add("step");
-      notes.push(
-        <>
-          <span className="promql-code promql-keyword">step()</span> resolves to
-          the query step (the interval between data points in a range query, or
-          the default evaluation step for instant queries).
-        </>
-      );
-    }
-    return;
-  }
-  if (node.op === "range") {
-    if (!seen.has("range")) {
-      seen.add("range");
-      notes.push(
-        <>
-          <span className="promql-code promql-keyword">range()</span> resolves to
-          the query range (the total time window covered by a range query).
-        </>
-      );
-    }
-    return;
-  }
-  if ((node.op === "min_of" || node.op === "max_of") && node.lhs && node.rhs) {
-    const key = `${node.op}:${formatDurationNode(node.lhs)}:${formatDurationNode(node.rhs)}`;
-    if (!seen.has(key)) {
-      seen.add(key);
-      notes.push(
-        <>
-          <span className="promql-code promql-keyword">{node.op}()</span> returns
-          the {node.op === "min_of" ? "smaller" : "larger"} of{" "}
-          <span className="promql-code">{formatDurationNode(node.lhs)}</span> and{" "}
-          <span className="promql-code">{formatDurationNode(node.rhs)}</span>.
-        </>
-      );
-    }
-  }
-  collectNotes(node.lhs, notes, seen);
-  collectNotes(node.rhs, notes, seen);
-};
-
-const durationExprNote = (expr: DurationNode): ReactNode => {
-  const notes: ReactNode[] = [];
-  collectNotes(expr, notes, new Set<string>());
-  if (notes.length === 0) {
-    return null;
-  }
-  const exprCode = (
-    <span className="promql-code">{formatDurationNode(expr)}</span>
-  );
-  if (notes.length === 1) {
-    return (
-      <Text fz="sm" mt="sm">
-        In the duration expression {exprCode}, {notes[0]}
-      </Text>
-    );
-  }
-  return (
-    <Text fz="sm" mt="sm">
-      In the duration expression {exprCode}:
-      <List mt="xs" withPadding>
-        {notes.map((note, i) => (
-          <List.Item key={i}>{note}</List.Item>
-        ))}
-      </List>
-    </Text>
-  );
-};
 
 interface ExplainViewProps {
   node: ASTNode | null;
